Move activitypub serialization into a module
