+++ fedireads/activitypub/__init__.py
@@ -1,0 +1,5 @@
+''' bring activitypub functions into the namespace '''
+from .actor import get_actor
+from .collection import get_add, get_remove
+from .create import get_create
+from .status import get_review, get_status
